Extract StoryHeader
